requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dataclasses-json==0.5.7
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.1
frozenlist==1.3.3
gpt-index==0.5.18
gptcache==0.1.15
greenlet==2.0.2
gunicorn==20.1.0
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.142
llama-index==0.5.17.post1
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.3
pycryptodomex==3.17
pydantic==1.10.5
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
pytz==2022.7.1
PyYAML==6.0
rake-nltk==1.0.6
regex==2022.10.31
requests==2.28.2
sentencepiece==0.1.97
six==1.16.0
SQLAlchemy==1.4.46
tenacity==8.2.2
tiktoken==0.3.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing-extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.3
XlsxWriter==3.0.8
yarl==1.8.2
zipp==3.15.0
